requirements.txt
alembic==1.5.5
Cerberus==1.3.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
